pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jsonschema>=4.24.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
